refactor: migrate package name from openforhatenabookmark to openforhatebu
- Move source files to jp.sane.openforhatebu package
- Update package declarations in all Kotlin files
- Update databinding imports
- Update test package name assertions
